refactor: short 타입 필드를 int로 통일 및 불필요한 캐스팅 제거 (#246)

+++ src/main/java/com/campusnavi/backend/community/post/entity/PostImage.java
@@ -22,9 +22,9 @@ public class PostImage{
     private String imageKey;
 
     @Column(nullable = false)
-    private short sortOrder = 0;
+    private int sortOrder = 0;
 
-    public static PostImage create(Post post, String imageKey, short sortOrder) {
+    public static PostImage create(Post post, String imageKey, int sortOrder) {
         PostImage image = new PostImage();
         image.post = post;
         image.imageKey = imageKey;
