+++ src/main/java/com/campusnavi/backend/official/post/repository/OfficialPostQueryRepository.java
@@ -152,7 +152,7 @@ public class OfficialPostQueryRepository {
                 .join(officialPostAiMeta.tag, tag)
                 .leftJoin(att).on(att.post.id.eq(officialPost.id)
                         .and(att.isImage.isTrue())
-                        .and(att.sortOrder.eq((short) 0)))
+                        .and(att.sortOrder.eq(0)))
                 .where(baseCondition(condition));
     }
 }
